feat(lua): update nvim-treesitter configuration
This commit updates the configuration for nvim-treesitter in the `lua/dikka/init.lua` file. The `ensure_installed` option now includes additional language parsers for markdown, javascript, go, rust, typescript, c, lua, vim, and vimdoc. The `highlight` option is enabled and the `additional_vim_regex_highlighting` option is set to false.

Additionally, a new keymap `<leader>z` is added to the configuration, which sets the `foldmethod` option to `'expr'` and the `foldexpr` option to `'nvim_treesitter#foldexpr()'` to load folds.